Move prod types to production
