+++ BTCTrackerTests/Presentation/ExchangePresenterTests.swift
@@ -8,80 +8,6 @@
 import XCTest
 import BTCTracker
 
-struct ExchangeViewModel: Equatable, Hashable {
-
-    private let exchange: Exchange
-
-    init(exchange: Exchange) {
-        self.exchange = exchange
-    }
-
-    var message: String {
-        return exchange.symbol + " is worth " + String(format: "$%0.2f", exchange.rate)
-    }
-
-}
-
-protocol ExchangeLoadingView {
-    func display(isLoading: Bool)
-}
-
-protocol ExchangeErrorView {
-    func display(error: String?)
-}
-
-protocol ExchangeView {
-    func display(viewModel: ExchangeViewModel)
-}
-
-final class ExchangePresenter {
-
-    private let loadingView: ExchangeLoadingView
-    private let errorView: ExchangeErrorView
-    private let exchangeView: ExchangeView
-    private let mapper: (Exchange) -> ExchangeViewModel
-    private let currentDate: () -> Date
-    private var lastUpdated: Date?
-    private let dateFormatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.dateStyle = .medium
-        formatter.timeStyle = .short
-        return formatter
-    }()
-
-    private var failureMessage: String {
-        var message = "Failed to update value."
-        if let lastUpdated = lastUpdated {
-            message += " Showing last updated value from \(dateFormatter.string(from: lastUpdated))"
-        }
-        return message
-    }
-
-    init(exchangeView: ExchangeView, loadingView: ExchangeLoadingView, errorView: ExchangeErrorView, mapper: @escaping (Exchange) -> ExchangeViewModel, currentDate: @escaping () -> Date) {
-        self.exchangeView = exchangeView
-        self.loadingView = loadingView
-        self.errorView = errorView
-        self.mapper = mapper
-        self.currentDate = currentDate
-    }
-
-    func didStartLoading() {
-        loadingView.display(isLoading: true)
-    }
-
-    func didFinishLoading(with exchange: Exchange) {
-        errorView.display(error: nil)
-        loadingView.display(isLoading: false)
-        exchangeView.display(viewModel: mapper(exchange))
-        lastUpdated = currentDate()
-    }
-
-    func didFinishLoading(with error: Error) {
-        loadingView.display(isLoading: false)
-        errorView.display(error: failureMessage)
-    }
-}
-
 final class ExchangePresenterTests: XCTestCase {
 
     func test_init_doesNotSendMessagesToView() {
